Update TransactionsTable.tsx
Following your help request on Discord. The component was not responsive because there was a conflict of classes between the TableHeads and TableCells.
The TableHeads for Channel and Category are responsive to md devices while the correspondent TableCells only the Category is responsive.
To solve that, all those Heads and Cells need to have the same classes for responsiveness.
To an even better result, if you change from medium (md) devices to large (lg) devices the table will respond better.

This is my analysis after testing your code on my end.